requirements.txt
spotipy
lyricsgenius
openai
pymongo
fastapi[all]
graphene
spaCy
textblob
gensim
wordcloud
transformers
torch
requests
pandas
joblib
uvicorn
pytest
python-decouple